requirements.txt
kafka-python
pandas
scikit-learn
